FAC-372: Remove CheckAssessmentExistencePort and move checks to adapters

+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/out/persistence/assessment/AssessmentPersistenceJpaAdapter.java
@@ -21,7 +21,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-import static org.flickit.assessment.core.common.ErrorMessageKey.GET_ASSESSMENT_PROGRESS_ASSESSMENT_RESULT_NOT_FOUND;
+import static org.flickit.assessment.core.common.ErrorMessageKey.*;
 import static org.flickit.assessment.data.jpa.core.assessment.AssessmentJpaEntity.Fields.ASSESSMENT_KIT_ID;
 import static org.flickit.assessment.data.jpa.core.assessment.AssessmentJpaEntity.Fields.SPACE_ID;
 import static org.springframework.data.jpa.domain.Specification.where;
@@ -35,7 +35,6 @@ public class AssessmentPersistenceJpaAdapter implements
     GetAssessmentProgressPort,
     GetAssessmentPort,
     DeleteAssessmentPort,
-    CheckAssessmentExistencePort,
     CountAssessmentsPort,
     CheckUserAssessmentAccessPort {
 
@@ -68,6 +67,9 @@ public class AssessmentPersistenceJpaAdapter implements
 
     @Override
     public UpdateAssessmentPort.Result update(UpdateAssessmentPort.AllParam param) {
+        if (!repository.existsById(param.id()))
+            throw new ResourceNotFoundException(UPDATE_ASSESSMENT_ID_NOT_FOUND);
+
         repository.update(
             param.id(),
             param.title(),
@@ -95,12 +97,10 @@ public class AssessmentPersistenceJpaAdapter implements
 
     @Override
     public void deleteById(UUID id, Long deletionTime) {
-        repository.delete(id, deletionTime);
-    }
+        if (!repository.existsById(id))
+            throw new ResourceNotFoundException(DELETE_ASSESSMENT_ID_NOT_FOUND);
 
-    @Override
-    public boolean existsById(UUID id) {
-        return repository.existsByIdAndDeletedFalse(id);
+        repository.delete(id, deletionTime);
     }
 
     @Override
